+++ multiboot2/Changelog.md
@@ -6,6 +6,9 @@
   name tag. However, this might also be relevant for users of custom multiboot2
   tags that use DSTs as types. See the example provided in the doc of the
   `get_tag` method.
+- renamed `MULTIBOOT2_BOOTLOADER_MAGIC` to `MAGIC`
+- added a `builder` feature and a `builder` module with a `Multiboot2InformationBuilder`
+  struct
 
 ## 0.15.1 (2023-03-18)
 - **BREAKING** `MemoryMapTag::all_memory_areas()` was renamed to `memory_areas`